move out some higher level components into pages dir
