+++ src/pages/AustinClouds.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Map from './Map';
+import Map from '../components/Map';
 import Layers from '../layers'
 import shaders from '../layers/shaders'
